% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analysis of football injuries}
\usage{
analyze(injury, data)
}
\arguments{
\item{injury}{Kind of injury}

\item{data}{data.frame which contains the data of fbinjuries}
}
\value{
analyze returns an object of class "fbinjuries". This object contains data of sick days caused by injury.
}
\description{
Analysis of football injuries
}
\examples{
data(fbinjuries)
analyze(injury = "Kreuzbandriss", data = fbinjuries)
}

