% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitlnorm.R
\name{fitlnorm}
\alias{fitlnorm}
\title{Estimate the parameters of a log-lormal distribution}
\usage{
fitlnorm(x)
}
\arguments{
\item{x}{random variable}
}
\description{
Estimate the parameters of a log-lormal distribution
}

