% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fplayer.R
\name{fplayer}
\alias{fplayer}
\title{Find a set of players}
\usage{
fplayer(position, minsickdays, maxsickdays, data)
}
\arguments{
\item{position}{player position:

\code{position = "TW"} (goalkeeper);

\code{position = "AW"} (defense);

\code{position = "MF"} (midfield);

\code{position = "ST"} (attack)}

\item{minsickdays}{the number of minimal sickdays in the past}

\item{maxsickdays}{the number of maximal sickdays in the past}

\item{data}{data frame which contains the data of fbinjuries}
}
\description{
Find a set of players
}
\examples{
data(fbinjuries)
fplayer(position = "AW", minsickdays = 10, maxsickdays = 18, data = fbinjuries)
}

