% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPercentile.R
\name{getPercentile}
\alias{getPercentile}
\title{Return the percentile for the number of sick days caused by injury}
\usage{
getPercentile(days, injury, data)
}
\arguments{
\item{days}{number of sick days}

\item{injury}{name of the injury}

\item{data}{data frame which contains the data of fbinjuries}
}
\value{
return the percentile
}
\description{
Return the percentile for the number of sick days caused by injury
}
\examples{
data(fbinjuries)
getPercentile(days = 30, injury = "Muskelfaserriss", data = fbinjuries)
}

