% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mkfile.R
\name{mkfile}
\alias{mkfile}
\title{Creates a textfile containing R output in your current working directory}
\usage{
mkfile(x, filename, setdir)
}
\arguments{
\item{x}{file content}

\item{filename}{filename without extension}

\item{setdir}{set path for textfile (default: current working directory)}
}
\description{
Creates a textfile containing R output in your current working directory
}
\examples{
string <- "Hello World!"
mkfile(x = string, "my_file")

\dontrun{
mkfile(x = string, "my_file", setdir = "your_path")
}
}
\seealso{
\itemize{
  \item \code{\link{getwd}} for getting working directory
  \item \code{\link{setwd}} for setting working directory
}
}

