% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_fbinjuries.R
\name{summary.fbinjuries}
\alias{summary.fbinjuries}
\title{Summary-Method for an object of class "fbinjuries"}
\usage{
\method{summary}{fbinjuries}(x, ...)
}
\arguments{
\item{x}{an object of class "fbinjuries"}

\item{...}{...}
}
\description{
Summary-Method for an object of class "fbinjuries"
}

