% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyse of a injury}
\usage{
analyze(injury, data)
}
\arguments{
\item{injury}{injury of interest}

\item{data}{data.frame which contains the data of fbinjuries}
}
\value{
summary of sickdays caused by injury
}
\description{
Analyse of a injury
}

