% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analysis of sick days caused by football injuries}
\usage{
analyze(injury, data)
}
\arguments{
\item{injury}{Kind of injury}

\item{data}{data frame which contains the data of fbinjuries}
}
\value{
analyze returns an object of class "fbinjuries". This object contains data of sick days caused by injury.

\code{Minimum} minimum sick days caused by injury

\code{firstQuartile} 0.25-quantile of sick days caused by injury

\code{Median} median of sick days caused by injury

\code{Mean} mean sick days caused by injury

\code{thirdQuartile} 0.75-quantile of sick days caused by injury

\code{Maximum} maximum sick days caused by injury

\code{numberOfObservation} number of injury

\code{data} subset of data frame containing data from fbinjuries

\code{injury} name of injury
}
\description{
Analysis of sick days caused by football injuries
}
\examples{
data(fbinjuries)
analyze(injury = "Kreuzbandriss", data = fbinjuries)
}

