% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fplayer.R
\name{fplayer}
\alias{fplayer}
\title{Find players who can play a special position...}
\usage{
fplayer(position, sickdays, data)
}
\arguments{
\item{position}{player position; \code{position = "TW"} (goalkeeper);
\code{position = "AW"} (defense);
\code{position = "MF"} (midfield);
\code{position = "ST"} (attack)}

\item{sickdays}{the number of maximal sickdays in the past}

\item{data}{data.frame which contains the data of fbinjuries}
}
\description{
Find players who can play a special position...
}

