% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mkfile.R
\name{mkfile}
\alias{mkfile}
\title{Creates a textfile with a report in your current working directory.}
\usage{
mkfile(x, filename)
}
\arguments{
\item{x}{object of class "fbinjuries"}

\item{filename}{filename without extension.}
}
\description{
Creates a textfile with a report in your current working directory.
}

