% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_fbinjuries.R
\name{plot.fbinjuries}
\alias{plot.fbinjuries}
\title{Plot-Method for an object of class "fbinjuries"}
\usage{
\method{plot}{fbinjuries}(x, ...)
}
\arguments{
\item{x}{an object of class "fbinjuries"}

\item{...}{...}
}
\description{
Plot-Method for an object of class "fbinjuries"
}

