% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_fbinjuries.R
\name{print.fbinjuries}
\alias{print.fbinjuries}
\title{Print-Method for an object of class "fbinjuries"}
\usage{
\method{print}{fbinjuries}(x, ...)
}
\arguments{
\item{x}{an object of class "fbinjuries"}

\item{...}{Arguments to be passed to or from other methods}
}
\value{
print information
}
\description{
Print-Method for an object of class "fbinjuries"
}
\examples{
data(fbinjuries)
kreuzbandriss <- analyze(injury = "Kreuzbandriss", data = fbinjuries)
kreuzbandriss
}

