% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Creates a report of a player.}
\usage{
report(name, data)
}
\arguments{
\item{name}{name of the player.}

\item{data}{data.frame which contains the data of fbinjuries}
}
\value{
A list of interesting information.
}
\description{
Creates a report of a player.
}

