% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_fbinjuries.R
\name{summary.fbinjuries}
\alias{summary.fbinjuries}
\title{Summarize an object of class "fbinjuries"}
\usage{
\method{summary}{fbinjuries}(object, ...)
}
\arguments{
\item{object}{Object of class "fbinjuries"}

\item{...}{Arguments to be passed to or from other methods}
}
\description{
Summarize an object of class "fbinjuries"
}

